# Conflicts:
#	.idea/misc.xml
#	app/src/main/java/com/example/openlog/viewmodel/SharedViewModel.kt

+++ app/src/main/java/com/example/openlog/ui/LineGraph.java
@@ -13,17 +13,17 @@ import kotlin.Pair;
 
 public class LineGraph {
 
-    List<Pair<Integer, Date>> dataSet;
+    List<Pair<Float, Date>> dataSet;
     org.eazegraph.lib.charts.ValueLineChart cubicValueLineChart;
     ValueLineSeries series = new ValueLineSeries();
 
-    public LineGraph(List<Pair<Integer, Date>> dataSet, ValueLineChart cubicValueLineChart) {
+    public LineGraph(List<Pair<Float, Date>> dataSet, ValueLineChart cubicValueLineChart) {
         this.dataSet = dataSet;
         this.cubicValueLineChart = cubicValueLineChart;
         setValues(dataSet);
     }
 
-    public void setValues(List<Pair<Integer, Date>> newDataSet) {
+    public void setValues(List<Pair<Float, Date>> newDataSet) {
         cubicValueLineChart.clearChart();
         cubicValueLineChart.setShowIndicator(false);
         series.getSeries().clear();
@@ -36,7 +36,7 @@ public class LineGraph {
         if (dataSet.size() > 1) {
             series.setColor(0xFF56B7F1);
 
-            for (Pair<Integer, Date> pair : dataSet) {
+            for (Pair<Float, Date> pair : dataSet) {
                 series.addPoint(new ValueLinePoint(DateTimeFormatter.Companion.formatAsDate(pair.component2()), pair.component1().floatValue()));
             }
